refactor: DIP, OCP 원칙에 따른 리팩터링
